Delete tratamento de exceção para recurso não encontrado

+++ src/main/java/com/desafio5/dscommer_v3/services/ProductService.java
@@ -6,9 +6,11 @@ import com.desafio5.dscommer_v3.repositories.ProductRepository;
 import com.desafio5.dscommer_v3.services.exceptions.ResourceNotFoundException;
 import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.GetMapping;
 
@@ -18,11 +20,12 @@ import java.util.Optional;
 public class ProductService {
     @Autowired
     private ProductRepository repository;
+
     @Transactional(readOnly = true)
     public ProductDTO findById(Long id) {
         Optional<Product> result = repository.findById(id);
         //Product product = result.get();
-        Product product = result.orElseThrow(()-> new ResourceNotFoundException("Recurso não encontrado. Verifique o id"));
+        Product product = result.orElseThrow(() -> new ResourceNotFoundException("Recurso não encontrado. Verifique o id"));
         ProductDTO productDTO = new ProductDTO(product);
         return productDTO;
     }
@@ -32,27 +35,37 @@ public class ProductService {
         Page<Product> result = repository.findAll(pageable);
         return result.map(x -> new ProductDTO(x));
     }
+
     @Transactional
-    public ProductDTO insert(ProductDTO dto){
+    public ProductDTO insert(ProductDTO dto) {
         Product entity = new Product();
         copyDtoToEntity(dto, entity);
         entity = repository.save(entity);
         return new ProductDTO(entity);
     }
+
     @Transactional
-    public ProductDTO update(Long id, ProductDTO dto){
+    public ProductDTO update(Long id, ProductDTO dto) {
         try {
             Product entity = repository.getReferenceById(id);
             copyDtoToEntity(dto, entity);
             entity = repository.save(entity);
             return new ProductDTO(entity);
-        }catch (EntityNotFoundException e){
+        } catch (EntityNotFoundException e) {
             throw new ResourceNotFoundException("Updade não executado. Id não existe.");
         }
     }
-    @Transactional
-    public void delete(Long id){
-        repository.deleteById(id);
+
+    @Transactional(propagation = Propagation.SUPPORTS)
+    public void delete(Long id) {
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException("Delete não executado. Id não encontrado");
+        }
+        try {
+            repository.deleteById(id);
+        }catch (DataIntegrityViolationException e){
+           // throw new DatabaseException("falha de integridade referêncial");
+        }
     }
 
     private void copyDtoToEntity(ProductDTO dto, Product entity) {
